comparator/__init__.py
__all__ = [
    "levenshtein_distance",
    "distance_is_acceptable",
    "normalize",
]

from Levenshtein import distance as levenshtein_distance
from .comparator import (
    distance_is_acceptable,
    normalize,
)
