pyproject.toml
[tool.poetry]
name = "ozi"
version = "0.1.0"
description = ""
authors = ["Vsevolod Bazhan <vsevozhan@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
djangorestframework = "^3.11.1"
djangorestframework-camel-case = "^1.2.0"
python-Levenshtein = "^0.12.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
isort = "^5.4.2"
Faker = "^4.1.2"
coverage = "^5.2.1"
pytest = "^6.0.1"
pytest-django = "^3.9.0"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
